// blesong-angular/src/app/models/template.model.ts
export class TemplateModel {

  idx: string;
  template_name: string;
  template_type: string;
  template_group: string;
  template_head_font: string;
  template_head_fontsize: string;
  template_head_fontcolor: string;
  template_body_font: string;
  template_body_fontsize: number;
  template_body_fontcolor: string;
  template_footer_font: string;
  template_footer_fontsize: string;
  template_footer_fontcolor: string;
  template_animation: number;
  createdAt: number;
  updatedAt: number;


}